third_party/blink/web_tests/virtual/stable/external/wpt/performance-timeline/tentative/with-filter-options-originA-expected.txt

This is a testharness.js-based test.
[FAIL] GetEntries with filter options.
  assert_equals: navigationEntries should be 1. expected 1 but got 3
Harness: the test ran to completion.
